Aligned scoring and paths
