+++ .github/workflows/cli-scripts-deploy-mlcompute-create-with-user-identity.yml
@@ -1,0 +1,31 @@
+name: cli-scripts-deploy-mlcompute-create-with-user-identity
+on:
+  workflow_dispatch:
+  schedule:
+    - cron: "0 0/4 * * *"
+  pull_request:
+    branches:
+      - main
+    paths:
+      - cli/deploy-mlcompute-create-with-user-identity.sh
+      - .github/workflows/cli-scripts-deploy-mlcompute-create-with-user-identity.yml
+      - cli/setup.sh
+jobs:
+  build:
+    runs-on: ubuntu-latest
+    steps:
+    - name: check out repo
+      uses: actions/checkout@v2
+    - name: azure login
+      uses: azure/login@v1
+      with:
+        creds: ${{secrets.AZ_CREDS}}
+    - name: setup
+      run: bash setup.sh
+      working-directory: cli
+      continue-on-error: true
+    - name: scripts installs
+      run: sudo apt-get upgrade -y && sudo apt-get install uuid-runtime jq -y
+    - name: test script script
+      run: set -e; bash -x deploy-mlcompute-create-with-user-identity.sh
+      working-directory: cli
