+++ .github/workflows/cli-scripts-deploy-safe-rollout-k8s-online-endpoints.yml
@@ -2,25 +2,17 @@ name: cli-scripts-deploy-safe-rollout-k8s-online-endpoints
 on:
   workflow_dispatch:
   schedule:
-    - cron: "0 1 * * *"
+    - cron: "0 0/4 * * *"
   pull_request:
     branches:
       - main
     paths:
-      - cli/endpoints/online/**
       - cli/deploy-safe-rollout-k8s-online-endpoints.sh
       - .github/workflows/cli-scripts-deploy-safe-rollout-k8s-online-endpoints.yml
       - cli/setup.sh
-      - .github/amlarc-tool.sh
 jobs:
   build:
     runs-on: ubuntu-latest
-    env:
-      KEY_VAULT_NAME: amlarcgithubworkflowkv
-      OWNERS: amlarc@microsoft.com
-      GITHUB_REPO: https://github.com/Azure/azureml-examples
-      WORKFLOW_URL: https://github.com/Azure/azureml-examples/actions/workflows/cli-scripts-deploy-safe-rollout-k8s-online-endpoints.yml
-      TSG_ID: https://microsoft.sharepoint.com/teams/Vienna/SiteAssets/Vienna%20Notebook/AMLArcTSG.one#Troubleshooting%20github%20workflow&section-id=69e8bdb1-5734-4b07-967a-5a50a91cf040&page-id=53a1a232-f9f0-4192-b7d2-0474848ebb18&end
     steps:
     - name: check out repo
       uses: actions/checkout@v2
@@ -33,21 +25,7 @@ jobs:
       working-directory: cli
       continue-on-error: true
     - name: scripts installs
-      run: sudo apt-get upgrade -y && sudo apt-get install uuid-runtime jq -y && sudo apt-get install xmlstarlet
+      run: sudo apt-get upgrade -y && sudo apt-get install uuid-runtime jq -y
     - name: test script script
       run: set -e; bash -x deploy-safe-rollout-k8s-online-endpoints.sh
       working-directory: cli
-    - name: file IcM when fails
-      if: ${{ failure() && github.event_name == 'schedule' }}
-      run: |
-        # download certificates
-        export ICM_HOST_NAME=ICM-HOST-PROD
-        export ICM_CONNECTOR_ID_NAME=ICM-CONNECTOR-ID-PROD
-        export ICM_ROUTING_ID_NAME=ICM-ROUTING-ID-PROD
-        set -e; bash -x .github/amlarc-tool.sh download_icm_cert
-        export ICM_HOST=$(cat icm_host)
-        export CONNECTOR_ID=$(cat icm_connector_id)
-        export ROUTING_ID=$(cat icm_routing_id)
-        export SUMMARY=$(set -e; bash -x .github/amlarc-tool.sh gen_summary_for_github_test)
-        set -e; bash -x .github/amlarc-tool.sh file_icm
-      timeout-minutes: 30
