+++ cli/deploy-custom-container-triton.sh
@@ -9,21 +9,23 @@ export ACR_NAME="<CONTAINER_REGISTRY_NAME>"
 
 export ENDPOINT_NAME=endpt-moe-`echo $RANDOM`
 
-# TODO: DELETE
-export ACR_NAME="valwallaceskr" 
-export ENDPOINT_NAME="minimal-amlinfsrv"
-export ENDPOINT_NAME="endpoint-12343"
-
-# <set_base_path_and_copy_model>
-export BASE_PATH="endpoints/online/custom-container/triton"
-mkdir $BASE_PATH/models
-cp -r endpoints/online/triton/single-model/models $BASE_PATH
-cp -r endpoints/online/triton/single-model/triton_densenet_scoring.py $BASE_PATH
-cp -r endpoints/online/triton/single-model/densenet_labels.txt $BASE_PATH
-# </set_base_path_and_copy_model>
+# <set_base_path_and_copy_assets>
+export PARENT_PATH="endpoints/online/custom-container"
+export ASSET_PATH="endpoints/online/triton/single-model"
+export BASE_PATH="$PARENT_PATH/triton_context"
+rm -rf $BASE_PATH && mkdir -p $BASE_PATH/models $BASE_PATH/code
+cp -r $ASSET_PATH/models $BASE_PATH
+cp $ASSET_PATH/triton_cc_scoring.py $BASE_PATH/code/score.py
+cp $ASSET_PATH/densenet_labels.txt $BASE_PATH/code
+cp $PARENT_PATH/triton-cc-deployment.yaml $BASE_PATH/deployment.yaml
+cp $PARENT_PATH/triton-cc.dockerfile $BASE_PATH/Dockerfile
+sed -i "s/{{acr_name}}/$ACR_NAME/g;\
+        s/{{endpoint_name}}/$ENDPOINT_NAME/g;" $BASE_PATH/deployment.yaml
+curl -o $BASE_PATH/peacock.jpg https://aka.ms/peacock-pic 
+# </set_base_path_and_copy_assets>
 
 # <login_to_acr>
-az acr login -n ${ACR_NAME} 
+az acr login -n $ACR_NAME
 # </login_to_acr> 
 
 # TODO: Delete
@@ -37,30 +39,25 @@ az acr build -t azureml-examples/triton-cc:latest -r $ACR_NAME $BASE_PATH
 az ml online-endpoint create -n $ENDPOINT_NAME --auth-mode key
 # </create_endpoint>
 
-cp $BASE_PATH/deployment.yaml $BASE_PATH/deployment_original.yaml
-sed -i "s/ACR_NAME/$ACR_NAME/" $BASE_PATH/deployment.yaml
-
 # <create_deployment>
 az ml online-deployment create --endpoint-name $ENDPOINT_NAME -f $BASE_PATH/deployment.yaml --all-traffic
 # </create_deployment> 
 
+# Get accessToken
+echo "Getting access token..."
+TOKEN=$(az ml online-endpoint get-credentials -n $ENDPOINT_NAME --query accessToken -o tsv)
+
+# Get scoring url
+echo "Getting scoring url..."
+SCORING_URL=$(az ml online-endpoint show -n $ENDPOINT_NAME --query scoring_uri -o tsv)
+echo "Scoring url is $SCORING_URL"
+
 # <test_online_endpoint>
-az ml online-endpoint invoke -n ${ENDPOINT_NAME} --request-file "$BASE_PATH/model-1/sample-request.json"
+curl -d "@$BASE_PATH/peacock.jpg" -H "Content-Type: image/jpeg" -H "Authorization: {Bearer $TOKEN}" $SCORING_URL
 # </test_online_endpoint>
- 
-curl -X GET https://aka.ms/peacock-pic -O
-
-curl -X POST  localhost:5001/score -data-file "@peacock-pic"  -H "Content-Type: image/jpeg"
-
-curl -X GET localhost:5001/score
 
 # <delete_online_endpoint>
 az ml online-endpoint delete -y -n $ENDPOINT_NAME
 # </delete_online_endpoint>
 
-# <delete_environment>
-az ml environment archive -y -n minimal-inf-cc-env
-# </delete_environment>
-
-rm -rf $BASE_PATH/model-1
-rm $BASE_PATH/deployment.yaml && mv $BASE_PATH/deployment_original.yaml $BASE_PATH/deployment.yaml 
+rm -rf $BASE_PATH
