+++ cli/endpoints/online/custom-container/triton-cc-deployment.yaml
@@ -1,0 +1,18 @@
+$schema: https://azuremlschemas.azureedge.net/latest/managedOnlineDeployment.schema.json
+name: triton-cc-deployment
+endpoint_name: {{endpoint_name}}
+environment: 
+  name: triton-cc-env
+  image: {{acr_name}}.azurecr.io/azureml-examples/triton-cc:latest
+  inference_config:
+      liveness_route:
+          path: /
+          port: 5001
+      readiness_route:
+          path: /
+          port: 5001
+      scoring_route:
+          path: /score
+          port: 5001
+instance_type: Standard_DS2_v2
+instance_count: 1
