--- cli/endpoints/online/custom-container/triton/deployment.yaml
@@ -1,18 +1,0 @@
-$schema: https://azuremlschemas.azureedge.net/latest/managedOnlineDeployment.schema.json
-name: green
-endpoint_name: endpoint-12343 
-environment: 
-  name: triton_test
-  image: valwallaceskr.azurecr.io/azureml-examples/triton-cc:latest
-  inference_config:
-    liveness_route:
-      path: /
-      port: 8000
-    readiness_route:
-      path: /
-      port: 8001
-    scoring_route:
-      path: /model_1/1
-      port: 8002
-instance_type: Standard_DS2_v2 #Standard_NC6s_v3
-instance_count: 1
