--- cli/endpoints/online/custom-container/triton/requirements.txt
@@ -1,3 +1,0 @@
-numpy
-pillow
-numpy
